refactor(anchor-navigation): put sub-components in folders
