+++ src/components/anchor-navigation/anchor-navigation-item/anchor-navigation-item.d.ts
