+++ src/components/anchor-navigation/anchor-section-divider/anchor-section-divider.style.js
@@ -1,0 +1,16 @@
+import styled from "styled-components";
+
+import { baseTheme } from "../../../style/themes";
+
+const StyledAnchorDivider = styled.div.attrs({
+  "data-element": "anchor-navigation-divider",
+})`
+  background-color: ${({ theme }) => theme.anchorNavigation.divider};
+  height: 1px;
+`;
+
+StyledAnchorDivider.defaultProps = {
+  theme: baseTheme,
+};
+
+export default StyledAnchorDivider;
