+++ src/components/anchor-navigation/index.d.ts
@@ -1,2 +1,3 @@
-export { default as AnchorNavigation } from './anchor-navigation';
-export { default as AnchorNavigationItem } from './anchor-navigation-item';
+export { default as AnchorNavigation } from "./anchor-navigation";
+export { default as AnchorNavigationItem } from "./anchor-navigation-item/anchor-navigation-item";
+export { default as AnchorSectionDivider } from "./anchor-section-divider/anchor-section-divider";
